Dispensing Update
Made the rfDisp dispensing controller more robust by restructuring the
input parameters and the way they are parsed.  Will likely need to
expand again with the inclusion of the syringe pump.
Made it possible to pass in a struct with plate (and dispensing)
properties to initialize the GUI to a predefined set of parameters.